streamlit_app/requirements.txt
streamlit>=1.30.0
google-cloud-bigquery[pandas,pyarrow]>=3.20.0
google-auth>=2.20.0
plotly>=5.18.0
pandas>=2.0.0
wordcloud>=1.9.0
matplotlib>=3.8.0
db-dtypes>=1.1.0
